Removing io.formats.common and moving functions to formats.format
